move code to new dir structure
