+++ go.mod
@@ -12,6 +12,8 @@ require (
 	github.com/onsi/gomega v1.9.0
 	github.com/stretchr/testify v1.5.1
 	go.uber.org/zap v1.16.0
+	golang.org/x/lint v0.0.0-20200302205851-738671d3881b // indirect
+	golang.org/x/tools v0.0.0-20201201192219-a1b87a1c0de4 // indirect
 	k8s.io/api v0.18.2
 	k8s.io/apiextensions-apiserver v0.18.2
 	k8s.io/apimachinery v0.18.2
