+++ operator/internal/component/component_test.go
